requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
Flask-Limiter>=3.8.0
# Backend: USE_TORCH=1 w server.py wymusza PyTorch w transformers (omija TensorFlow).
# Jeśli nadal błąd DLL TensorFlow: pip uninstall -y tensorflow tensorflow-intel tensorboard
sentence-transformers<3.0.0,>=2.7.0
# PyTorch - install separately with: pip install torch==2.2.2 torchvision==0.17.2 --index-url https://download.pytorch.org/whl/cpu
# torch==2.2.2
# torchvision==0.17.2
numpy>=1.26.0
faiss-cpu>=1.8.0
google-generativeai>=0.7.0
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
wikipedia-api>=0.13.0
pytest>=8.0.0
pytest-cov>=5.0.0
redis>=5.0.0
